+++ CHANGES.md
@@ -1,3 +1,4 @@
 ## Optimizations:
 
 - Instead of dumping the whole candidate list, add it to the graph as soon as it is recieved.
+- Implement a dedicated sessions manager